requirements.txt
numpy
pandas
flask
flask_caching
redis
google-api-python-client==1.12.8
python-dotenv
